wiplayit_app/requirements.txt
asgiref==3.2.7
Babel==2.8.0
certifi==2019.11.28
chardet==3.0.4
coverage==5.0.4
defusedxml==0.6.0
Django==3.0.4
django-allauth==0.41.0
django-cors-headers==3.2.1
django-guardian==2.2.0
django-js-asset==1.2.2
django-mptt==0.11.0
django-phonenumber-field==4.0.0
django-rest-auth==0.9.5
django-twilio==0.13.0.post1
django-webpack-loader==0.7.0
djangorestframework==3.11.0
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==4.4.0
idna==2.9
mysqlclient==1.4.6
oauthlib==3.1.0
phonenumbers==8.12.4
Pillow==7.0.0
PyJWT==1.7.1
python3-openid==3.1.0
pytz==2019.3
requests==2.23.0
requests-oauthlib==1.3.0
six==1.14.0
sqlparse==0.3.1
twilio==6.40.0
urllib3==1.25.8
python-dotenv
